refactor: promise list response 수정
