+++ src/main/java/com/owl/payrit/domain/promise/repository/PromiseRepository.java
@@ -1,4 +1,4 @@
-package com.owl.payrit.domain.promise.reposiroty;
+package com.owl.payrit.domain.promise.repository;
 
 import com.owl.payrit.domain.member.entity.Member;
 import com.owl.payrit.domain.promise.entity.Promise;
@@ -9,4 +9,5 @@ import java.util.List;
 public interface PromiseRepository extends JpaRepository<Promise, Long> {
 
     List<Promise> findAllByWriter(Member writer);
+    List<Promise> findAllByOwner(Member owner);
 }
